(PC-30403)[PRO] refactor: move specific audience wordings in collective or individual actions bar
