+++ pro/src/pages/Offers/Offers/OffersActionsBar/utils/computeActivationSuccessMessage.ts
@@ -1,0 +1,7 @@
+export const computeActivationSuccessMessage = (nbSelectedOffers: number) => {
+  const successMessage =
+    nbSelectedOffers > 1
+      ? 'offres ont bien été publiées'
+      : 'offre a bien été publiée'
+  return `${nbSelectedOffers} ${successMessage}`
+}
